Test:add UT for matrix::reshape (#2117)

+++ source/module_base/matrix.cpp
@@ -353,7 +353,7 @@ double matrix::trace_on(void) const
 // }
 
 // Peize Lin add 2017-05-27
-void matrix::reshape( const double nr_new, const double nc_new, const bool flag_zero )
+void matrix::reshape( const int nr_new, const int nc_new, const bool flag_zero )
 {
 	assert( nr*nc == nr_new*nc_new );
 	nr=nr_new;
