+++ source/module_base/matrix.h
@@ -70,7 +70,7 @@ public:
 
 	void get_extreme_eigen_values(double &ev_lower, double &ev_upper) const;	// mohan add 2011-01-13
 
-	void reshape( const double nr_new, const double nc_new,  const bool flag_zero = true );		// Peize Lin add 2017-05-27
+	void reshape( const int nr_new, const int nc_new,  const bool flag_zero = true );		// Peize Lin add 2017-05-27
 
 	double max() const;					// Peize Lin add 2016-09-08
 	double min() const;					// Peize Lin add 2016-09-08
